backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2  # Add explicit bcrypt version
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
python-dateutil==2.8.2
apscheduler==3.10.4
webauthn>=2.0.0
requests==2.31.0